requirements.txt
Django==4.2.4
Pillow==10.0.0
requests==2.31.0
django-memoize==2.3.1
